% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumor_covariates.R
\name{clr_transform}
\alias{clr_transform}
\title{Centered log-ratio transform for compositional fractions}
\usage{
clr_transform(fractions, pseudo = 0.001)
}
\arguments{
\item{fractions}{Samples x cell types matrix of non-negative fractions.}

\item{pseudo}{Pseudocount replacing zeros.}
}
\description{
Cell-type fractions sum to a constant, so use CLR before regression.
}
